[jsscripting] Fix setTimeout/setInterval fails if delay is not provided or null
According to the [mdn docs](https://developer.mozilla.org/en-US/docs/Web/API/Window/setTimeout), the delay parameter for setTimeout and setInterval is optional and may be omitted.
Previously, such method calls failed as the undefined or null value was passed to the Java layer, which does not understand those values.
This is fixed by now handling such values in the JS polyfills, making sure a valid number is always passed to the Java layer.

Signed-off-by: Florian Hotze <dev@florianhotze.com>