[ecoflow] Fix missing updates for PowerStream channels (#19632)
A 'quota' (update) MQTT message looks like this:
{ "cmdId":1,
  "cmdFunc":20,
  "addr":"iot",
  "params":{ "plugTotalWatts":0,
             "invOutputWatts":1000,
             [...]
           }
}

Previously, for the PowerStream, the 'params' object was named 'param'.
It is unclear why exactly the change happened (cloud server change? PS
firmware change?), but it seems to have happened on 2025-10-03.

Adjust to the change and log a warning instead of crashing if such a
case ever happens again.

Signed-off-by: Danny Baumann <dannybaumann@web.de>