Update SMHI Binding to support TimeSeries and migrate to new API
SMHI have updated their API, which have led to changes in the parameter names, which are now more descriptive. The parameter names are used as channel IDs, which leads to those being updated as well. In addition, some new parameters have been added, and one enum-type parameter have changed the mapping.

At the same time, support for TimeSeries have been introduced, so the full forecast can be sent to a single Item.

Signed-off-by: Anders Alfredsson <andersb86@gmail.com>