[matter] Support Switchbot Candle Warmer (#19699)
The SwitchBot candle warmer is in violation of the matter spec and says its a extended color device when its not, this causes us to ignore channels as we assume there's a ColorControlCluster since thats mandatory per the spec.  This also adds a few more device types to our color/switch/Level check.

Signed-off-by: Dan Cunningham <dan@digitaldan.com>