[homeassistant] Update python packages
This was missed when bringing the rest of the Python code up to date
with current Home Assistant.

Also commit the lockfile, which is a new feature thanks to GraalPy 25.

Also fix error handling error in Python template code: no functional
change, the logging should just give an actually helper message
instead of a NameError

Signed-off-by: Cody Cutrer <cody@cutrer.us>